# [xaml] improve performance in debug-mode