@StephaneDelcroix 2024-03-28
IIRC, there was a reason using GetFields() instead of GetField(). Could be that one looks up the hierarchy and the other not, or the order, or different versions using different profiles (netstandard1, ...), But it should be caught by unit tests